requirements.txt
flask
pyyaml
openai
Pillow
imagehash
tqdm
ffmpeg-python